requirements.txt
pillow
tensorflow
pandas
numpy
matplotlib
